LPS-93905 Move AssetRenderers from *.asset to *.asset.model packages
