+++ modules/apps/blogs/blogs-web/src/main/java/com/liferay/blogs/web/asset/model/BlogsEntryAssetRenderer.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.blogs.web.asset;
+package com.liferay.blogs.web.asset.model;
 
 import com.liferay.asset.display.page.portlet.AssetDisplayPageFriendlyURLProvider;
 import com.liferay.asset.kernel.model.AssetRendererFactory;
@@ -20,6 +20,7 @@ import com.liferay.asset.kernel.model.BaseJSPAssetRenderer;
 import com.liferay.asset.util.AssetHelper;
 import com.liferay.blogs.constants.BlogsPortletKeys;
 import com.liferay.blogs.model.BlogsEntry;
+import com.liferay.blogs.web.asset.BlogsEntryAssetRendererFactory;
 import com.liferay.blogs.web.internal.security.permission.resource.BlogsEntryPermission;
 import com.liferay.blogs.web.internal.util.BlogsEntryUtil;
 import com.liferay.portal.kernel.exception.PortalException;
