+++ modules/apps/calendar/calendar-web/src/main/java/com/liferay/calendar/web/internal/asset/CalendarBookingAssetRendererFactory.java
@@ -23,6 +23,7 @@ import com.liferay.calendar.model.Calendar;
 import com.liferay.calendar.model.CalendarBooking;
 import com.liferay.calendar.model.CalendarResource;
 import com.liferay.calendar.service.CalendarBookingLocalService;
+import com.liferay.calendar.web.internal.asset.model.CalendarBookingAssetRenderer;
 import com.liferay.calendar.web.internal.util.CalendarResourceUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
