+++ modules/apps/comment/comment-web/src/main/java/com/liferay/comment/web/internal/asset/CommentAssetRendererFactory.java
@@ -18,6 +18,7 @@ import com.liferay.asset.kernel.model.AssetRenderer;
 import com.liferay.asset.kernel.model.AssetRendererFactory;
 import com.liferay.asset.kernel.model.BaseAssetRendererFactory;
 import com.liferay.comment.constants.CommentConstants;
+import com.liferay.comment.web.internal.asset.model.CommentAssetRenderer;
 import com.liferay.comment.web.internal.constants.CommentPortletKeys;
 import com.liferay.portal.kernel.comment.Comment;
 import com.liferay.portal.kernel.comment.CommentManager;
