+++ modules/apps/contacts/contacts-web/src/main/java/com/liferay/contacts/web/internal/asset/UserAssetRendererFactory.java
@@ -18,6 +18,7 @@ import com.liferay.asset.kernel.model.AssetRenderer;
 import com.liferay.asset.kernel.model.AssetRendererFactory;
 import com.liferay.asset.kernel.model.BaseAssetRendererFactory;
 import com.liferay.contacts.constants.ContactsPortletKeys;
+import com.liferay.contacts.web.internal.asset.model.UserAssetRenderer;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.User;
