+++ modules/apps/document-library/document-library-web/src/main/java/com/liferay/document/library/web/internal/asset/DLFileEntryAssetRendererFactory.java
@@ -28,6 +28,7 @@ import com.liferay.document.library.kernel.service.DLAppLocalService;
 import com.liferay.document.library.kernel.service.DLFileEntryLocalService;
 import com.liferay.document.library.kernel.service.DLFileEntryTypeLocalService;
 import com.liferay.document.library.util.DLURLHelper;
+import com.liferay.document.library.web.internal.asset.model.DLFileEntryAssetRenderer;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.log.Log;
