+++ modules/apps/site/site-admin-web/src/main/java/com/liferay/site/admin/web/internal/asset/SiteAssetRendererFactory.java
@@ -20,6 +20,7 @@ import com.liferay.asset.kernel.model.BaseAssetRendererFactory;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.service.GroupLocalService;
+import com.liferay.site.admin.web.internal.asset.model.SiteAssetRenderer;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
