+++ modules/apps/tasks/tasks-portlet/docroot/WEB-INF/src/com/liferay/tasks/asset/model/TasksEntryAssetRenderer.java
@@ -15,7 +15,7 @@
  * Liferay Social Office. If not, see http://www.gnu.org/licenses/agpl-3.0.html.
  */
 
-package com.liferay.tasks.asset;
+package com.liferay.tasks.asset.model;
 
 import com.liferay.asset.kernel.model.BaseJSPAssetRenderer;
 import com.liferay.portal.kernel.model.User;
