+++ modules/apps/wiki/wiki-web/src/main/java/com/liferay/wiki/web/internal/trash/WikiPageTrashHandler.java
@@ -46,7 +46,7 @@ import com.liferay.wiki.model.WikiPageResource;
 import com.liferay.wiki.service.WikiPageLocalService;
 import com.liferay.wiki.service.WikiPageResourceLocalService;
 import com.liferay.wiki.service.WikiPageService;
-import com.liferay.wiki.web.internal.asset.WikiPageAssetRenderer;
+import com.liferay.wiki.web.internal.asset.model.WikiPageAssetRenderer;
 import com.liferay.wiki.web.internal.util.WikiPageAttachmentsUtil;
 
 import java.util.ArrayList;
